requirements.txt
Flask
tensorflow
pillow
numpy
sklearn
pickle
streamlit
